Move `vec_slice_shaped()` to `slice.R` and merge with `vec_slice()`
